Refactor modals: move form and textfield components from modal directory
